requirements.txt
# Core ML
torch>=2.0
torchvision>=0.15
opencv-python>=4.8
numpy>=1.24
Pillow>=10.0
onnx>=1.14

# VLM
transformers>=4.40
peft>=0.10
bitsandbytes>=0.43
accelerate>=0.30
qwen-vl-utils

# API
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9

# Testing
pytest>=8.0
httpx>=0.27
